requirements.txt
Django==2.0.7

celery<4.0.0 # Do not upgrade celery version since django-celery does not support celery 4.0 or later.
celery-haystack==0.10
coverage==4.5.4  # Updated from 4.5.1
ipython==7.8.0  # Updated from 6.5.0
jsmin==2.2.2
Pillow==6.1.0  # Updated from 5.2.0
piprot==0.9.10
premailer==3.6.1  # Updated from 3.2.0
psycopg2-binary==2.8.3  # Updated from 2.7.5
raven==6.9.0
requests==2.22.0  # Updated from 2.19.1
sorl-thumbnail==12.3 # Do not upgrade sorl-thumbnail version since django-sorl-thumbnail does not support sorl-thumbnail 12.4.1 or later.
suds-py3==1.3.3.0
unicodecsv==0.14.1
tblib==1.3.2
whoosh==2.7.4
xlwt==1.3.0

django-autoslug-iplweb==1.9.4
django-axes==4.4.0
django-braces==1.13.0
django-celery-transactions==0.3.6
django-celery==3.3.1  # Updated from 3.2.2
django-compressor==2.3  # Updated from 2.2
django-constance==2.4.0  # Updated from 2.3.1
django-crispy-forms==1.7.2
django-debug-toolbar==2.0  # Updated from 1.10.1
django-dynamic-fixture==2.0.0
django-extensions==2.2.1  # Updated from 2.1.2
django-haystack==2.8.1
django-markdown-deux==1.0.5
django-recaptcha==1.4.0
django-redis-cache==2.0.0  # Updated from 1.7.1
django-test-plus==1.3.1  # Updated from 1.1.1
djangorestframework==3.8.2
django-widget-tweaks==1.4.5
